sampledeploy/cdk.out/asset.806712f32eccce919eb5b2129b888a14c40561691f6606e64ac3656a334c548e/requirements.txt
boto3==1.34.19
botocore==1.34.19
ffmpeg-python==0.2.0
openai==1.65.4
python-dotenv==1.0.1
requests==2.31.0 